refactor(trajets): extract TrajetRow + _Chip to keep trajets_tab.dart under 400 lines
The 400-line file_length guard (#1680) flagged trajets_tab.dart at
415 lines after the #2018 landscape layout work. Splits the
trajet row + chip widget into its own file
(`lib/features/consumption/presentation/widgets/trajet_row.dart`)
under 150 lines. Same widget contract — the inline `_TrajetRow`
becomes the public `TrajetRow`.
